+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Added custom quality gates for password security which can be changed in setting
 * Complete reworking of the storage management for entries and details.
 * Stored details and entries are converted the first time the app is opened with data from before version 3.3.0.
 * Toggling biometrics for login requires confirming identity through biometrics.
+* Disabeling app login requires biometric authentication (if available) or confirmation through app password.
 
 <br>